Add job for image-diff plugin
This also adds a job template to build UI-only plugins. These are
plugins only providing html and js files, but no jar file. The
default plugin build job was not usable for these artifacts, since
it artifact archiver was expecting a jar-file and would fail, if
it was missing.

+++ jenkins/gerrit-bazel-build-plugin.sh
@@ -23,7 +23,10 @@ bazelisk build --spawn_strategy=standalone --genrule_strategy=standalone $TARGET
 bazelisk test --test_env DOCKER_HOST=$DOCKER_HOST //tools/bzl:always_pass_test plugins/{name}/...
 
 JAR="bazel-bin/plugins/{name}/{name}.jar"
-PLUGIN_VERSION=$(git describe  --always origin/{branch})
-echo -e "Implementation-Version: $PLUGIN_VERSION" > MANIFEST.MF
-jar ufm $JAR MANIFEST.MF && rm MANIFEST.MF
-echo "$PLUGIN_VERSION" > bazel-bin/plugins/{name}/$(basename $JAR-version)
+if test -f $JAR
+then
+  PLUGIN_VERSION=$(git describe  --always origin/{branch})
+  echo -e "Implementation-Version: $PLUGIN_VERSION" > MANIFEST.MF
+  jar ufm $JAR MANIFEST.MF && rm MANIFEST.MF
+  echo "$PLUGIN_VERSION" > bazel-bin/plugins/{name}/$(basename $JAR-version)
+fi
